Create AppointmentParticipation object as parent for Appointment object
