+++ src/main/java/nl/mrwouter/zermelo4j/appointments/Appointment.java
@@ -6,150 +6,24 @@ import java.util.stream.Collectors;
 import com.google.gson.JsonObject;
 
 /**
- * Appointment object for easy overview
+ * Appointment object
  */
-public class Appointment {
+public class Appointment extends AppointmentParticipation {
 
-    private final long id;
-    private final long start;
-    private final long end;
-    private final List<String> subjects;
-    private final List<String> teachers;
-    private final List<String> groups;
-    private final List<String> locations;
-    private final String remark;
-    private final String changeDescription;
-    private final String startTimeSlot;
-    private final String endTimeSlot;
-    private final Boolean isParticipation;
-    private final Boolean valid;
-    private final Boolean cancelled;
-    private final Boolean modified;
-    private final Boolean moved;
-    private final Boolean isNew;
-    private final AppointmentType appointmentType;
+    private final boolean valid;
+    private final boolean modified;
+    private final boolean moved;
+    private final boolean isNew;
 
-    public Appointment(boolean isParticipation, long id, long start, long end, String startTimeSlot, String endTimeSlot,
+    public Appointment(long id, long start, long end, String startTimeSlot, String endTimeSlot,
                        List<String> subjects, List<String> teachers, List<String> groups, List<String> locations,
-                       AppointmentType appointmentType, String remark, Boolean valid, Boolean cancelled, Boolean modified,
-                       Boolean moved, Boolean isNew, String changeDescription) {
-        this.isParticipation = isParticipation;
-        this.id = id;
-        this.start = start;
-        this.end = end;
-        this.startTimeSlot = startTimeSlot;
-        this.endTimeSlot = endTimeSlot;
-        this.subjects = subjects;
-        this.teachers = teachers;
-        this.groups = groups;
-        this.locations = locations;
-        this.appointmentType = appointmentType;
-        this.remark = remark;
+                       AppointmentType appointmentType, String remark, boolean valid, boolean cancelled, boolean modified,
+                       boolean moved, boolean isNew, String changeDescription) {
+        super(id, start, end, startTimeSlot, endTimeSlot, subjects, teachers, groups, locations, appointmentType, remark, cancelled, changeDescription);
         this.valid = valid;
-        this.cancelled = cancelled;
         this.modified = modified;
         this.moved = moved;
         this.isNew = isNew;
-        this.changeDescription = changeDescription;
-    }
-
-    /**
-     * Get the ID of this appointment
-     *
-     * @return id of this appointment
-     */
-    public long getId() {
-        return id;
-    }
-
-    /**
-     * Get the appointment type
-     *
-     * @return appointment type
-     */
-    public AppointmentType getType() {
-        return appointmentType;
-    }
-
-    /**
-     * Get the starting timeslot of this appointment. Returns null when not provided.
-     *
-     * @return start time slot
-     */
-    public String getStartTimeSlot() {
-        return startTimeSlot;
-    }
-
-    /**
-     * Get the ending timeslot of this appointment. Returns null when not provided.
-     *
-     * @return end time slot
-     */
-    public String getEndTimeSlot() {
-        return endTimeSlot;
-    }
-
-    /**
-     * Get the time when this lesson starts (Seconds since epoch)
-     *
-     * @return time when this appointment starts
-     */
-    public long getStart() {
-        return start;
-    }
-
-    /**
-     * Get the time when this appointment starts (Seconds since epoch)
-     *
-     * @return time when this appointment ends
-     */
-    public long getEnd() {
-        return end;
-    }
-
-    /**
-     * Get a list of subjects for this appointment
-     *
-     * @return array of subjects
-     */
-    public List<String> getSubjects() {
-        return subjects;
-    }
-
-    /**
-     * Get a list of teachers for this appointment
-     *
-     * @return list of teachers
-     */
-    public List<String> getTeachers() {
-        return teachers;
-    }
-
-    /**
-     * Get a list of groups for this appointment
-     *
-     * @return list of groups
-     */
-    public List<String> getGroups() {
-        return groups;
-    }
-
-    /**
-     * Get a list of locations for this appointment
-     *
-     * @return list of locations
-     */
-    public List<String> getLocations() {
-        return locations;
-    }
-
-    /**
-     * Remark for this appointment (For example: "Don't forget your books")
-     *
-     * @return remark for this appointment
-     */
-    public String getRemark() {
-        return remark;
     }
 
     /**
@@ -159,48 +33,26 @@ public class Appointment {
      * possible.'
      *
      * @return true if appointment is valid
-     * @throws AppointmentParticipationException when using
-     *                                           getAppointmentParticipations
      */
-    public boolean isValid() throws AppointmentParticipationException {
-        if (isParticipation) {
-            throw new AppointmentParticipationException("Can't invoke isValid when using getAppointmentParticipations");
-        }
+    public boolean isValid() {
         return valid;
     }
 
     /**
-     * Check if appointment is cancelled
-     *
-     * @return true if appointment is cancelled
-     */
-    public boolean isCancelled() {
-        return cancelled;
-    }
-
-    /**
-     * Check if appointment is modified. When using the
-     * {@link nl.mrwouter.zermelo4j.ZermeloAPI#getAppointmentParticipations(String, int, int)}
-     * method, isModified is only a check if {@link #getChangeDescription()} is
-     * empty.
+     * Check if appointment is modified.
      *
      * @return true if appointment is modified
      */
     public boolean isModified() {
-        return isParticipation ? !changeDescription.isEmpty() : modified;
+        return modified;
     }
 
     /**
      * Check if the appointment has moved
      *
      * @return true if appointment has moved
-     * @throws AppointmentParticipationException when using
-     *                                           getAppointmentParticipations
      */
-    public boolean isMoved() throws AppointmentParticipationException {
-        if (isParticipation) {
-            throw new AppointmentParticipationException("Can't invoke isValid when using getAppointmentParticipations");
-        }
+    public boolean isMoved() {
         return moved;
     }
 
@@ -209,50 +61,21 @@ public class Appointment {
      * originally scheduled.'
      *
      * @return true if appointment is new
-     * @throws AppointmentParticipationException when using
-     *                                           getAppointmentParticipations
      */
-    public boolean isNew() throws AppointmentParticipationException {
-        if (isParticipation) {
-            throw new AppointmentParticipationException("Can't invoke isValid when using getAppointmentParticipations");
-        }
+    public boolean isNew() {
         return isNew;
     }
 
     /**
-     * Get the description of the change.
-     *
-     * @return textual description of the change.
-     */
-    public String getChangeDescription() {
-        return changeDescription;
-    }
-
-    /**
-     * Added custom #toString() method for easy debugging.
+     * {@inheritDoc}
      */
     @Override
-    public String toString() {
-        JsonObject appointment = new JsonObject();
-        appointment.addProperty("id", id);
-        appointment.addProperty("start", start);
-        appointment.addProperty("end", end);
-        appointment.addProperty("startTimeSlot", startTimeSlot);
-
-        appointment.addProperty("endTimeSlot", endTimeSlot);
-        appointment.addProperty("teachers", String.join(",", teachers));
-        appointment.addProperty("subjects", String.join(",", subjects));
-        appointment.addProperty("groups", String.join(",", groups));
-        appointment.addProperty("locations", String.join(",", locations));
-        appointment.addProperty("appointmentType", appointmentType.name());
-        appointment.addProperty("remark", remark);
-        appointment.addProperty("valid", (!isParticipation ? valid : null));
-        appointment.addProperty("cancelled", cancelled);
-        appointment.addProperty("modified", (!isParticipation ? modified : null));
-        appointment.addProperty("moved", (!isParticipation ? moved : null));
-        appointment.addProperty("new", (!isParticipation ? isNew : null));
-        appointment.addProperty("changeDescription", changeDescription);
-
-        return appointment.toString();
+    public JsonObject toJson() {
+        JsonObject appointment = super.toJson();
+        appointment.addProperty("valid", this.isValid());
+        appointment.addProperty("modified", this.isModified());
+        appointment.addProperty("moved", this.isMoved());
+        appointment.addProperty("new", this.isNew());
+        return appointment;
     }
 }
