+++ src/main/java/nl/mrwouter/zermelo4j/appointments/AppointmentComparator.java
@@ -5,10 +5,10 @@ import java.util.Comparator;
 /**
  * Comparator used to make sure that all appointments are sorted from start to beginning.
  */
-public class AppointmentComparator implements Comparator<Appointment> {
+public class AppointmentComparator implements Comparator<AppointmentParticipation> {
 
     @Override
-    public int compare(Appointment app1, Appointment app2) {
+    public int compare(AppointmentParticipation app1, AppointmentParticipation app2) {
         return Long.compare(app1.getStart(), app2.getStart());
     }
 }
